requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openai>=1.0.0
anthropic>=0.7.0
pillow>=10.0.0
python-dotenv>=1.0.0
keyring
markitdown>=0.0.1a2
dashscope>=1.14.0
openpyxl>=3.0.0
python-docx>=1.2.0
tavily-python>=0.3.0